masterdevice_raspi_ver/rasberry_UI/rasberry_UI/requirements.txt
dash==3.2.0
dash-bootstrap-components==2.0.3
pandas==2.3.2
numpy==2.2.6
plotly==6.3.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.4
